refactor(sqlite3): move live performQuery into the Rails-layout file
The unwired `performQuery` in `sqlite3/database-statements.ts` was a port
written against better-sqlite3's native sync API (raw connection, spread
binds) with no callers, while the adapter ran its own private `_performQuery`
against the async driver abstraction. Two implementations of one Rails method;
api:compare's `perform_query` coverage pointed at the dead one.

Rewrite the exported `performQuery` against the async `SqliteStatement` driver
abstraction so it IS the live primitive, and have `AbstractSQLite3Adapter`
delegate to it. Keeps the statement pool (`_cachedStatement`),
`_maybeEnableReadBigInts`, and the `stmt.reader` / separate affected-rows read
(incl. the `INSERT ... RETURNING` reader===true case). api:compare now points
at reachable code; the three Rails calls the converged body deliberately omits
(`prepare`, `Result.new`, `empty?`) are recorded in the wide-calls exclude.
