+++ scripts/api-compare/call-mismatches-wide-exclude.json
@@ -1,5 +1,26 @@
 [
   {
+    "package": "activerecord",
+    "tsFile": "connection-adapters/sqlite3/database-statements.ts",
+    "rubyName": "perform_query",
+    "call": "prepare",
+    "reason": "Satisfied by a different path: the async driver abstraction prepares through the statement pool via `this._cachedStatement(sql)` (which caches / falls back to a fresh prepare) rather than calling `raw_connection.prepare` directly."
+  },
+  {
+    "package": "activerecord",
+    "tsFile": "connection-adapters/sqlite3/database-statements.ts",
+    "rubyName": "perform_query",
+    "call": "new",
+    "reason": "Deliberate deviation (documented in performQuery): reads return raw row hashes and writes source the affected-rows / insert-rowid atomically from the RunResult, so no `ActiveRecord::Result.new(columns, to_a)` is constructed here."
+  },
+  {
+    "package": "activerecord",
+    "tsFile": "connection-adapters/sqlite3/database-statements.ts",
+    "rubyName": "perform_query",
+    "call": "empty?",
+    "reason": "Deliberate deviation: the branch axis is `stmt.reader && !isWrite` (not `column_count.zero?`) and the no-row branch returns `[]` directly rather than `ActiveRecord::Result.empty`, so `binds.empty?` / empty-Result construction is not mirrored."
+  },
+  {
     "package": "abstractcontroller",
     "tsFile": "base.ts",
     "rubyName": "_valid_action_name?",
